requirements.txt
python-dotenv==1.0.0
simple-term-menu==1.6.6
openai==1.57.2